requirements.txt
mcp>=1.2.0
uvicorn>=0.27.0
starlette>=0.36.0
aiohttp>=3.9.3
sseclient-py>=1.8.0
requests>=2.31.0 